Attach Lo-Dash as _ to the REPL context
The `_` character is special in the Node REPL. See http://nodejs.org/api/repl.html#repl_repl_features

So that means after doing something like > 1 + 2, which will produce 3, the value of _ is 3

See https://github.com/joyent/node/blobb5175003bca45e23bd20d9559096dafda3e0ad7c/lib/repl.js#L318

Fixed the issue by wrapping the ["line" callback](https://github.com/joyent/node/blob/b5175003bca45e23bd20d9559096dafda3e0ad7c/lib/repl.js#L244) with a function which keeps track of the current state of _ attached to the server context.
When the original line callback is executed the currVal will be whatever expression result it set _ to and the prevVal will be what it was before that. So if a user defined _ with their own value that value will be in the prevVal otherwise the prevVal will be lodash. So we assign currVal the prevVal value after execution of the line callback to bypass the line callbacks default behavior.

Big thanks to John-David Dalton!

References:

Closes https://github.com/borisdiakur/n_/issues/1